Allow unlock during rollback
During usm rollback if the user lock the host that's not the next one
to be deployed he will not be able to unlock it and also not be able
to do a deploy host-rollback either. This commit fix this issue by
doing a request to an usm endpoint to check if the host to be unlocked
is the next host to be rollback.

Test Plan:
PASS: Lock/Unlock controller-1 after software deploy abort with both
controllers in deploy-host-rollback-pending state.
PASS: AIO-DX Lock/unlock of both controllers in non deployment scenario.
PASS: AIO-DX Lock/unlock of both controllers during deployment.
PASS: AIO-SX Lock/unlock

Story: 2010676
Task: 50981

Depends-On: https://review.opendev.org/c/starlingx/update/+/928493

Change-Id: I64e362f64c0af33af57b64ed192c525d97656a3b
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>