Improve adding OSD for ceph bare-metal
This change is necessary because the timeout related to deferred
configs is getting the timestamp of the item when it was appended
to the list and not the timestamp when it starts applying the puppet.
Example: The config was added to the list at 08:30 and the config
started applying at 08:50, the timeout check will get 08:30 and
not 08:50
To avoid the timeout error, we can update the timestamp for a deferred
config with the same classes in each agent report using a new timestamp.

Test Plan:
 - PASS: Install a Standard system with 10 OSDs added at the same time

Closes-Bug: 2077120

Change-Id: Ib5e056ae3938db270ac2889667782af073e8a0e4
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>