Improvements to application dependency behavior on deploy activate
This change adds improvements that were discussed in the review
https://review.opendev.org/c/starlingx/config/+/927528.

An unnecessary check has been removed from the update_in_series
function in the upgrade script 65. Spelling corrections have also
been added.

These changes also resolve a bug for Zuul that runs the
"sysinv tox" command. The command was failing due to the import
"from sysinv.api.controllers.v1 import kube_app" which in turn
failed to resolve the import of the fm_core package which is not
found if imported directly.

This is an intermediate solution, but a new task will be created
to deal with this, resolving the issue with the Fault repository
and adjusting the tox of all applications.

Test Plan:
PASS: build-pkgs && build-image
PASS: run "tox -e flake8"
PASS: run "tox -e metadata" in metrics-server repository
PASS: run "tox -e metadata" in istio repository
PASS: Upgrade from stx-9 to master
PASS: 65-k8s-app-upgrade.sh script runs without errors
PASS: Applications that can be updated in parallel are updated
      in parallel.
PASS: Applications that need to be serially updated are serially
      updated.

Closes-bug: 2078948
Closes-bug: 2080063

Change-Id: I79c6e5e46ec3f4b693f1325d0241b29c60ec489e
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>