Update dnsmasq when subcloud's system-controller-oam pool is modified
If there is a need to update the system-controller-oam pool due to
a system-controller's OAM floating address change, it was not
happening at runtime. This will lead to lost of contact with
registry.central domain.

This review adds a runtime call to update dnsmasq, no need to
lock/unlock the subcloud to update the value.

Test Plan
=========

[PASS] Install an AIO-SX subcloud
[PASS] Modify System-Controller OAM floating address and in the
       subcloud modify system-controller-oam's addrpool
       - test that registry.central remains reachable with:
         - ping
         - docker pull

Closes-Bug: 2080506

Change-Id: I2c542301956cd06fb9cce98b24b1c85f0319a97e
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>