Add OSDs without creating a new config_uuid for rook-ceph
This change includes generating the config_uuid only
when the runtime_manifest is True.
This change will prevent an out-of-date alarm creation
when the backend is ceph-rook.

The test was run only on rook-ceph, because the only
change was to modify when the config_uuid is generated.
It will run on ceph-bare metal because it uses
runtime_manifest as True and will not run on rook-ceph
because it uses runtine_manifest as False

The Rook-Ceph configures OSDs using the application,
so no puppet is needed, and no need to create a
new config_uuid.

Test Plan:
 - PASS: Install a Duplex system, add the OSDs in all hosts,
         apply the rook-ceph app and check if the alarm list
         is empty

Story: 2011117
Task: 51006

Change-Id: Id0de8565ed62134904c22d838cba7b8c3e99f687
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>