Make test_kubeapi_health wrapper/decorator function
This change adds the test_kubeapi_health function/decorator to the
common folder and makes its access more generic. Previously it was
only available inside the 41-create-required-platform-certs.py
script.

Is expected that kube-apiserver responds within fractions of one
second (responsiveness < 100ms 95% of the time). This adjustment
is generous and adds a timeout for the request of a maximum of 5
seconds.

Test Plan:
PASS: Upgrade from stx-9 to master
PASS: Decorator added to the create_platform_certificates function
      of the upgrade script 41 and the check of the K8s APIs is
      performed successfully.

Closes-bug: 2080637

Change-Id: I76920ea2b4e78b2fe3b05dddb0f7f01b7a36816e
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>