Correct interface addr-mode config in mixed dual/single-stack nets
It was observed that in a scenario where two networks (e.g. management
as dual-stack and cluster-host as IPv4-only) are set on the same
interface, the interface's address mode configuration was incorrectly
set to "static" for both IPv4 and IPv6 when the interface-network API
was called. This caused the controller-X address allocated in the
other network (IPv6 in this case) to be removed.

For example, if management is dual-stack (IPv4,IPv6) and cluster-host
is IPv4-only, when the interface-network for cluster-host was created,
it removed the controller-0-mgmt address for IPv6.

This issue was caused by the incorrect address mode configuration when
the interface-network API was called. It was configuring to static for
both IPv4 and IPv6 regardless if the network was dual-stack or not.
This triggered a wipe-out of addresses in the other stack on the
particular interface.

Test Plan:
==========

Check that
 - all pools contain the expected addresses allocated
 - interface address mode are set to static in the affected protocol

[PASS] AIO-DX install with
        - mgmt (prim:IPv6,sec:IPv4)
        - cluster-host,pod,svc (prim:IPv6)
        - oam (prim:IPv6,sec:IPv4)
        - admin (prim:IPv6)
[PASS] AIO-DX install with
        - mgmt (prim:IPv6,sec:IPv4)
        - cluster-host,pod,svc (prim:IPv6,sec:IPv4)
        - oam (prim:IPv6,sec:IPv4)
        - admin (prim:IPv6,sec:IPv4)
[PASS] AIO-DX install with
        - mgmt (prim:IPv6)
        - cluster-host,pod,svc (prim:IPv6,sec:IPv4)
        - oam (prim:IPv6,sec:IPv4)
        - admin (prim:IPv6)
[PASS] AIO-DX install with
        - mgmt (prim:IPv6)
        - cluster-host,pod,svc (prim:IPv6)
        - oam (prim:IPv6)
        - admin (prim:IPv6)
[PASS] Using the single-stack installation, convert system to
       dual-stack.

Story: 2011027
Task: 51028

Change-Id: Id4837ed0d17a608b558070ed37874b95ce3c30a6
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>