CertificateRequest not deleted for a deleted host
This commit adds functions to handle with The CertificateRequests
that is used when a host is updating the certificates and when a
host is deleted from the system. First, a function verify if there
is a CertificateRequest already created and delete it if it's true.

Test Plan:
PASS: Full build, system install, bootstrap and unlock DX system w/
      a worker node added succesfully and all the nodes w/ unlocked
      enabled available status.
PASS: First, execute "kubectl get cr -n deployment" and observe a
      CertificateRequest related to all nodes. Then, execute
      "system host-delete <worker-hostname>" in the active controller
      to delete the node and observe the previously created
      CertificateRequest was deleted as well.

Story: 2010940
Task: 51042

Change-Id: Ia4a2d69e958f956879802c5ea8a345b5e99becb5
Signed-off-by: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>