Fix HTTPS enable for USM upgrade
During upgrade, in some scenarios where cert-mon has sequential
timeouts, it sometimes causes the system to not switch to HTTPS
immediately (cert-mon eventually recovers from the timeouts, and
still tries to install the 'ssl' certificate), or even to not
switch indefinitely (the cert-mon call reaches sysinv and
installs the 'ssl' cert, but when cert-mon recovers, it considers
the certificate already installed and doesn't proceed to enable
HTTPS).

In the setups where the timeouts were occurring, the certificate
installs were taking ~45s.

This review increases the timeouts for certificate update calls
from cert-mon to sysinv from 30 to 60 seconds. It also adds an
extra call to sysinv, to ensure that the patch to 'https_enabled'
is done even if cert-mon is still seeing timeouts.

Test plan:
PASS: With AIO-SX with HTTPS disabled, perform stx 9 to stx 10
      upgrade. Observe that #41 script is executed correctly
      and HTTPS is enabled during the execution.

Story: 2009811
Task: 51072

Change-Id: If43e48f68b725ae4f9d7b2e30ed5469837dc59ff
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>