Prevent error when pod conditions are unavailable
In scenarios where Kubernetes is unhealthy, pod conditions are not
reported and a NoneType error is raised when checking for readiness
states. This commit prevents the error by checking whether Kubernetes
reported pod conditions prior to evaluating any values.

The fix implemented here for the "check_pod_running_and_ready_probe"
method makes use of the same logic implemented for the
"check_pod_completed" method on
https://review.opendev.org/c/starlingx/config/+/927811. The common
portion of the logic was generalized and moved to a new method in order
to improve the reuse of that code.

Test plan:
PASS: build-pkgs && build-image
PASS: AIO-SX fresh install
PASS: AIO-SX upgrade from previous release

Closes-bug: 2082021

Change-Id: I3864b8752766519efe6f916b4760c436069dadbb
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>