Add message to IPsec client to confirm op completed
This commit add a message to the end of the execution of IPsec client
in order to confirm to IPsec server the end of the operation. This
change is implemented in order to fix the scenario where IPsec server
sets the configured host to IPsec enabled without the node being
completely configured.

Test Plan:
PASS: Deploy a standard system and upgrade from stx 9 to stx 10.
      Observe that IPsec is configured and enabled on hosts, keys and
      certificates are correctly generated, SAs are established and
      nodes are reachable. Observe on ipsec-auth.log that operation 1
      is completed to each node on the environment.
PASS: On a standard system, observe on ipsec-auth.log that IPsec client
      generated the last message after puppet apply is complete for op
      code 1.
PASS: Manually execute operation 2 and 3 for IPsec client and observe
      that operations succeeds.

Story: 2010940
Task: 51111

Change-Id: I27af8f31aad90bee06b842ffd60f8446d619011a
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>