Fix misleading log message when VIM progress status is undefined
An exception is thrown in the check_nodes_stable function when
vim_progress_status is null for controller hosts, causing a warning log
message to be generated. The message is not clear about what caused the
issue and leads to unnecessary confusion. This change adds a check to
make sure no exception is thrown when the variable is null.

Test plan

Test is covered by included automated test function.
AIO-DX installation was performed for sanity.

Closes-Bug: 1915864
Change-Id: I751dd78ca4b5980b5682768a68c968dc269d44f1
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>