Rename/move upgrade scripts
Moved the scripts bellow to be grouped together:
- remove-psp-service-parameter.py
- delete-encrypted-fs-service-parameter.sh
- add-tls-version-cipher-parameters.sh
- oidc-service-parameters-migration.py

These scripts make changes to service parameters in K8s API, related
to security features. Since the API restarts on each apply, it's
better to execute the scripts and perform a single apply after the
last script.

However, some of the parameter changes need to be done at the
beginning of the upgrade, so the scripts were placed following the
executed earlier amongst them (remove-psp-service-parameter.py), which
was placed in position #02, so the scripts will occupy the positions
02-05.

An extra script will be added in a follow up change, after 'oidc-
service-parameters-migration.py' to perform the service-parameter
apply and monitor the restart of K8s API, in the position #06.

The other scripts were renamed to accommodate that, keeping their
order of execution between themselves. Also, some scripts that had the
same number were renamed to have a unique number.

Test plan:
PASS: Perform upgrade activation from stx 8 -> stx 10.

Story: 2010676
Task: 51103

Change-Id: I6aa2ae2a9d41793bab3a9ea67b9218b11cbbbd1d
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>