Fix ceph mon creation and update
Ceph can not have monitor on workers when the system is duplex, so the
check about adding monitors on workers was improved. And the host-fs
update was not correct because it was updating only the last host-fs
ceph in the host-fs list.

This issue only occurs for ceph bare-metal

Test Plan:
 - PASS: Install ceph with monitor in SX/DX/DX+/STD
 - PASS: Increase ceph mon storage in SX/DX/DX+/STD
 - PASS: Check if the host-fs ceph is changed with ceph mon in DX/DX+

Closes-bug: 2083333

Change-Id: I8db7a79085945c0c7b0fb6607d522c0972c7a5e8
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>