Retain address after removing interface associated to the network
When an interface is assigned to a network, the address associated with
that interface is updated with the interface ID information. When the
interface is removed, it must disassociate the interface ID from the
address entry and not delete the entry itself.

Test plan:
==========

[PASS] Create an admin network and associate it with a VLAN interface in
a SX system mode. Then, remove the VLAN interface. The admin controller
address of the host must be retained in address table.

[PASS] Create an admin network and associate it with a VLAN interface in
a SX system mode. Then, disassociate the VLAN interface from the admin
network. The admin controller address of the host must be retained in
address table.

[PASS] Create an admin network and associate it with a VLAN interface in
subcloud SX system mode. Then, remove the VLAN interface. The admin
controller address of the host must be retained in address table.

[PASS] Create an admin network and associate it with a VLAN interface in
subcloud SX system mode. Then, disassociate the VLAN interface from the
admin network. The admin controller address of the host must be retained
in address table.

[PASS] Create a storage network. Then, create a VLAN interface linked to
the storage network. After upgrading from the previous version to the
current version one, delete the VLAN interface. The controller storage
address must be retained in the address database table.

Story: 2010722
Task: 51119

Change-Id: I3474215cd51a54db1d2722e43cc16237445d44d2
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>