K8s upgrade: Remove the deprecated flag pod-eviction-timeout
The pod-eviction-timeout flag of the kube-controller-manager has
been deprecated and completely removed in K8s v1.27. This removal
was already addressed in STX 9 in the following review:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/895628

However, the pod-eviction-timeout flag is present in the
kube-controller-manager for K8s versions in STX 8 and continues to
be available after the platform upgrades from STX 8 to other
releases. This presence of the flag has resulted in pods getting
stuck in the pending state after upgrading K8s versions.

This change ensures that the pod-eviction-timeout flag is removed
from the controller manager in K8s v1.27 during upgrades, resolving
the issue with pending pods.

Test Plan:
PASS: Perform the platform upgrade from STX 8 to STX 10 for AIO-SX,
      upgrade K8s from v1.24 to v1.29, and verify all the pods.
PASS: Install the ISO with K8s v1.24 on AIO-SX, upgrade to v1.29,
      and verify all the pods.

Closes-Bug: 2083383

Change-Id: I10f69ec048944c2c961f957b4125f29aceee977e
Signed-off-by: Ramesh Kumar Sivanandam <rameshkumar.sivanandam@windriver.com>