Fix mon state of workers in the database
When a monitor is created on a worker, and that worker is still
LOCKED, it is not possible to execute the puppet manifest runtime,
which makes it impossible to update the database.

To resolve this, when a monitor is created on a LOCKED host,
the initial state in the database is 'configuring-on-unlock'.
After unlocking, it is checked whether the host is “available”,
and then the mon status is changed to “configured”.

Test Plan:
- PASS: Fresh install on STD with DM and check
        output of 'system ceph-mon-list'
- PASS: Move monitor from worker-0 to worker-1

Closes-Bug: 2083558

Change-Id: I40ed513118b2a9b5414907c565b56a3d8adb6962
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>