Sync system-local-ca in subcloud upgrade
In legacy upgrade, system-local-ca was synced between SystemController
and subcloud either by dcmanager or a ansible playbook, depending on
the system type. With USM, the secret remains as it was before the
upgrade, which causes issues if the secret is Opaque.

Subclouds will have the secret as Opaque if the user hasn't performed
the cert-manager migration procedure in the subcloud. This is not
supported anymore, as from stx 10 onward the REST API / GUI and Local
Docker registry certificates will be issued by default on the subcloud,
requiring system-local-ca's secret to be of TLS type.

This review adds an upgrade activation script that will execute in the
subcloud and recreate system-local-ca if necessary, retrieving the data
from the SystemController. It will attempt to connect using SSH thru
OAM or MGMT networks, retrying between them in case of failures.

Test plan:
PASS: Perform stx 8 -> stx 10 upgrade activation in a subcloud with
      system-local-ca of type Opaque. Verify that the Opaque secret is
      replaced by a TLS secret with the same data as the
      SystemController.
PASS: Perform stx 8 -> stx 10 upgrade activation in a subcloud with
      system-local-ca of owned by a Certificate object. Verify that
      the Certificate is deleted and a TLS secret with the same data as
      the SystemController is created.
PASS: Perform stx 8 -> stx 10 upgrade activation in a subcloud with
      system-local-ca of type Opaque. Switch of the OAM network.
      Verify that the Opaque secret is replaced by a TLS secret with
      the same data as the SystemController using the MGMT network.

Story: 2009811
Task: 51133

Change-Id: I97b05918b80223add692a8e09790a08cfe1fe0a2
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>