Fix OAM IP address configuration
The system oam-modify command does not include a parameter to
modify/update the range. For instance, when the oam-subnet is changed
from a /24 prefix to a /23 prefix through system oam-modify command,
the range retains the old end OAM IP values from the /24 prefix. A new
OAM IP address that falls within the second part of the network will be
out of range and lead to an invalid configuration. Although the command
is deprecated, the fix updates the end OAM IP from range according to
the configured subnet when system oam-modify command is used.
Additionally, for the address pool create/modify command, the system
currently permits IP configurations that fall within the subnet but are
outside the configured range. This update introduces validation to
ensure that IP addresses are within the configured range when the
address pool is created or modified.

Test Plan
=========

[PASS] Create an OAM address pool by specifying a range. The
configuration should complete successfully:
system addrpool-add --floating-address 10.20.10.125
--gateway-address 10.20.10.1 --ranges 10.20.10.1-10.20.10.254
teste 10.20.10.0 23

[PASS] Create an OAM address pool without specifying a range. The
configuration should complete successfully:
system addrpool-add --floating-address 10.20.10.125
--gateway-address 10.20.10.1  teste 10.20.10.0 23

[PASS] Create an address pool with a specific range and a floating IP
address that is outside the configured range but within the configured
subnet. The configuration should fail:
system addrpool-add --floating-address 10.20.11.125
--gateway-address 10.20.10.1 --ranges 10.20.10.1-10.20.10.254
teste 10.20.10.0 23

[PASS] Create an address pool with a specific range and a floating IP
address that is outside the subnet. The configuration should fail:
system addrpool-add --floating-address 10.20.11.125
--gateway-address 10.20.10.1 --ranges 10.20.10.1-10.20.10.254
teste 10.20.10.0 24

[PASS] Execute the same tests as before, but using the address pool
modify command.

[PASS] System configured with OAM subnet /24. Change the OAM IP address
using the system oam-modify command with a network that has a prefix
length of /23, ensuring the IP address falls within the first part of
the subnet. The configuration should complete successfully.

[PASS] System configured with OAM subnet /24. Change the OAM IP address
using the system oam-modify command with a network that has a prefix
length of /23, ensuring the IP address falls within the second part of
the subnet. The configuration should complete successfully.

[PASS] System configured with OAM subnet /24. Change the OAM IP address
using the system oam-modify command with a network that has a prefix
length of /23 and configure IP address out of subnet. The configuration
should fail.

[PASS] Bootstrap using OAM IPv4/IPv6 address out of range. Ansible
should validate the configuration and fail.

Closes-Bug: 2084173

Change-Id: I652d7215efdda50f777982cbc1edc9163a5a2d00
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>