Workaround: Remove 500.200 Alarm for cdi-uploadserver-client-cert
After uploading and applying the KubeVirt app, a 500.200 alarm is
triggered for the cdi-uploadserver-client-cert certificate.
This is a known issue, discussed on GitHub
https://github.com/kubevirt/containerized-data-importer/issues/2755
and has been fixed in CDI version 1.60.

This change serves as a workaround to suppress the alarm for the
cdi-uploadserver-client-cert certificate until we upgrade to
CDI version 1.60. This enable the subsequent Kubernetes upgrade
to initiateand finish successfully.
Using Kubernetes watch to continuously monitor the cdi in a separate
thread

Test Plan:
PASS: Upload and apply the kubevirt-app, then verify that the
500.200 alarm for cdi-uploadserver-client-cert does not appear
after running the health status check.
PASS: Perform single and Multi k8s upgrade.

Partial-Bug: 2084142

Change-Id: I1e17950bcdfc421a1424120e2ea5261623e9204c
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>