Truncate app progress string before saving to DB.
Truncate the string that stores the installation progress of an
StarlingX application.

This fixes an issue that was leading to an exception while saving the
progress string to the database. During error handling scenarios that
string may contain a message from a previously raised exception, which
may be larger than the database limit for the corresponding column.
That was causing applications to be stuck at applying state rather
than being updated to the "apply-failed" state after an error.

Test plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX fresh install.
PASS: Apply metrics-server before platform-integ-apps is applied.
      Confirm that metrics-server was moved to apply-failed.
PASS: Wait for platform-integ-apps to be applied.
      upload/apply/remove/delete metrics-server.
      Confirm that no errors were raised.

Closes-bug: 2084285

Change-Id: Ib2b369dd6c8b215fa8d9ba12064e7ebef42b6fc1
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>