Registry Image List Backwards Compatibility
The filter-out-untagged arg was introduced recently. This is always
sent with api calls, and fails on older systems with an unknown
argument error. This commit makes the cgtsclient send the
filter-out-untagged arg only when the user provides it. This makes
the cgtsclient backwards compatible with systems before this arg
was introduced.

Testing
PASS: system registry-image-list --filter-out-untagged
      on a current system
PASS: system registry-image-list on a current system
PASS: system registry-image-list pointing the cgtsclient
      at an old system before this arg was introduced

Closes-Bug: 2084294
Change-Id: Icae4179f6d979b23df9378f8fb6ddeaef659aa7e
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>