The command "tox -e metadata" fails
These changes resolve a bug for Zuul that runs the "sysinv tox"
command. The command was failing due to the import
"from sysinv.api.controllers.v1 import kube_app" which in turn
failed to resolve the import of the fm_core package which is not
found if imported directly.

This is an intermediate solution, but a new task has been created
to deal with this, resolving the issue with the Fault repository
and adjusting the tox of all applications.

This adjustment was previously made by this review:
https://review.opendev.org/c/starlingx/config/+/928784, but due to
new modifications it was removed by mistake.

Test Plan:
PASS: build-pkgs && build-image
PASS: run "tox -e flake8"
PASS: run "tox -e metadata" in metrics-server repository
PASS: run "tox -e metadata" in istio repository
PASS: Upgrade from stx-9 to master

Closes-bug: 2080063

Change-Id: I81bf5f02a69125b9095a71053ed8a20f377ce679
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>