Fix concurrent calls to _generate_dnsmasq_hosts_file causing FS error
The _generate_dnsmasq_hosts_file function can be called simultaneously
by two different threads. When the thread in which the function is
called later tries to read the dnsmasq.hosts.temp file, it has already
been renamed to dnsmasq.hosts by the other thread, causing a 'file not
found' error. This can happen during subcloud deploy bootstrap. In rare
cases the function is called at the same time from configure_ihost and
_controller_config_active_apply which is a periodic task, causing the
deploy to fail.
This change adds a decorator to the function which ensures that the
calls from different threads are executed in sequence instead of
concurrently.

Test plan

[PASS] Included automated test function
[PASS] AIO-SX subcloud deploy
[PASS] AIO-DX fresh install

Closes-Bug: 2084480
Change-Id: I43ec95d8555c89f5d7069ad655059193d5330e8b
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>