Remove virtual interfaces from LLDP audit
When a pod is created, it configures the virtual interface, assigns an
IP address, and applies firewall rules. Once this configuration is
complete, the interface is set to UP. At the same time, a periodic task
known as LLDP Audit is executed by sysinv. This task retrieves all
interfaces, including virtual ones, to gather neighbor information via
LLDP. For interfaces still in the creation process and in the DOWN
state, the LLDP audit temporarily sets them to UP to perform the LLDP
operation. After the audit, the interface reverts to its previous
state, which is DOWN, regardless of whether the virtual interface
management pod has set it to UP in the meantime. It is important to
exclude virtual interfaces from the LLDP audit process.

Test Plan
=========

[PASS] Reproduce one of the SX provisioning steps by executing the LLDP
audit and pod creation simultaneously. This is not a deterministic
scenario during normal provisioning, but it may occur depending on when
the LLDP audit is executed. This will create a race condition between
the creation of Calico virtual interfaces and the LLDP audit. To
effectively replicate the issue, it is essential that the LLDP audit
interface scanning takes place precisely at the moment of pod creation,
specifically, between the creation of the virtual interface and its
transition to the UP state. To facilitate this simulation, the audit
interval for LLDP can be reduced to 1 second. At the same time,
stimulate the cyclic restart of all pods during the stage following the
bootstrap, before the first unlock.

Story: 2011106
Task: 51194

Change-Id: I6662afc2b499879b255fd52306a4ca07b7997519
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>