Fixes Armada pod deletion crash
In some tests, it was observed that the Armada pod was getting
stuck in the "terminating" state and the command "kubectl delete
po <pod_ip>" was not enough to get it out of this state.

To fix this, the --now flag was added to the command that
deletes the pod. To improve robustness, retry logic was added
and a timeout was defined for each call to the delete command.

Test Plan:
PASS: build-pkgs && build-image
PASS: upgrade from stx-8 to master
PASS: armada pods are successfully deleted during execution of
      script 01 in the activate process.

Closes-bug: 2084682

Signed-off-by: David Bastos <david.barbosabastos@windriver.com>
Change-Id: I6598bb69168c132a8b460cda4a56c8d53484ef76