Add new reapply trigger for storage-backend-modify
Adds a new app reapply trigger for modifications on storage backend.
This is required by rook ceph to follow the replication and
min_replication values present in the storage-backend in the inventory.

Also fixes the storage backend add and modify actions for ceph-rook
only, correcting the parsing of the min_replication when using open
deployment model.

Test Plan:
  - PASS: Verify the new trigger with Rook-Ceph application
  - PASS: Verify storage-backend-add and modify actions with
          variations on replication and min_replication parameters
  - PASS: Build iso and install a environment with all changes on SX
          and STD.

Story: 2011066
Task: 51193

Change-Id: Iba4ddaef11edc50a8c03857a9180efc2d31ed6ae
Signed-off-by: Caio Correa <caio.correa@windriver.com>