Private keys are being dumped to sysinv
When user bootstrap and unlock a DC system controller private keys
are being dumped to sysinv.log. There are some points in the manager
that is logging the entire runtime_config object, one of the
runtime_config is to the upload cert files, in that case the log is
showing the entire content of .pem files. The solution is to hide
file content before logging the runtime_config to sysinv.log.

Test Plan:
PASS: Deploy and bootstrap a DC
PASS: Wait when defered runtime config are logged to sysinv and check
if there's no more key dumped

Closes-Bug: 2085630

Change-Id: Ied97632ecb9076dd60d78d7576261355ef3988a2
Signed-off-by: Joaci Morais <Joaci.deMorais@windriver.com>