Immediately audit subclouds recieved through notification
When a subcloud becomes online, cert-mon will insert it in the queue
to be audited in case it's not in the queue already. Since offline
subclouds will also be audited, it can take a long time before
auditing the subcloud that just became online, as offline subclouds
will likely timeout. This commit introduces a new parameter 'priority'
that will be set to 0 when an audit request is received through
notification. In this case, the subcloud will be inserted at the start
of the queue to be audited right away.

There are two caveats of this solution:
- If the subcloud is already in the queue, it will be inserted again
  and will be audited two times, as it's not possible to remove a
  subcloud from the queue.
- If the initial audit fail and needs to be re-queued, it will go to
  the end of the queue. This is because it's not possible to requeue
  at the start of the queue with a delay.

For the last caveat, the number of subclouds being audited in parallel
was raised, so although it goes to the the end of the queue, the time
to audit the subcloud again is shorter.

Test plan:
  - PASS: Start cert-mon in an environment with a lot of subclouds
          turned off, so all audits are timing out. Turn on some
          subclouds and verify there's audited right away.

Closes-bug: 2085540

Change-Id: I0efaf490af422e778e4a996a9d493ccc0d47e82e
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>